@@ -8,29 +8,28 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.1.3",
-        "ext-pdo_sqlite": "*",
-        "async-aws/core": "^1.6",
-        "async-aws/sns": "^1.1",
-        "async-aws/sqs": "^1.3",
-        "barryvdh/laravel-cors": "^0.11.3",
-        "doctrine/dbal": "^2.10",
-        "fideloper/proxy": "^4.0",
-        "laravel/framework": "5.7.*",
-        "laravel/tinker": "^1.0",
+        "php": "^7.2.5",
+        "async-aws/core": "^1.10",
+        "async-aws/sns": "^1.2",
+        "async-aws/sqs": "^1.5",
+        "doctrine/dbal": "^2.13",
+        "fideloper/proxy": "^4.2",
+        "fruitcake/laravel-cors": "^1.0",
+        "guzzlehttp/guzzle": "^6.3",
+        "laravel/framework": "^7.0",
+        "laravel/tinker": "^2.0",
         "league/flysystem-aws-s3-v3": "^1.0",
         "maatwebsite/excel": "^3.1",
         "sentry/sentry-laravel": "^2.5",
         "spatie/enum": "^2.3",
-        "tymon/jwt-auth": "1.0.*"
+        "tymon/jwt-auth": "^1.0"
     },
     "require-dev": {
-        "beyondcode/laravel-dump-server": "^1.0",
-        "filp/whoops": "^2.0",
-        "fzaninotto/faker": "^1.4",
-        "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^2.0",
-        "phpunit/phpunit": "^7.0"
+        "facade/ignition": "^2.0",
+        "fzaninotto/faker": "^1.9.1",
+        "mockery/mockery": "^1.3.1",
+        "nunomaduro/collision": "^4.1",
+        "phpunit/phpunit": "^8.5"
     },
     "config": {
         "optimize-autoloader": true,
