@@ -8,6 +8,7 @@ use App\Http\Controllers\Controller;
 use App\Needs;
 use App\Agency;
 use App\Applicant;
+use App\Enums\ApplicantStatusEnum;
 use App\LogisticRequest;
 use App\FileUpload;
 use App\Imports\LogisticImport;
@@ -30,16 +31,16 @@ class LogisticRequestController extends Controller
     public function finalList(Request $request)
     {
         $syncSohLocation = \App\PoslogProduct::syncSohLocation();
-        $request->request->add(['verification_status' => Applicant::STATUS_VERIFIED]);
-        $request->request->add(['approval_status' => Applicant::STATUS_APPROVED]);
-        $request->request->add(['finalized_by' => Applicant::STATUS_FINALIZED]);
+        $request->merge(['verification_status' => ApplicantStatusEnum::verified()]);
+        $request->merge(['approval_status' => ApplicantStatusEnum::approved()]);
+        $request->merge(['finalized_by' => ApplicantStatusEnum::finalized()]);
         // Cut Off Logistic Data
         $cutOffDateTimeState = \Carbon\Carbon::createFromFormat(config('wmsjabar.cut_off_format'), config('wmsjabar.cut_off_datetime'))->toDateTimeString();
         $cutOffDateTime = $request->input('cut_off_datetime', $cutOffDateTimeState);
         $today = \Carbon\Carbon::now()->toDateTimeString();
 
-        $request->request->add(['start_date' => $cutOffDateTime]);
-        $request->request->add(['end_date' => $today]);
+        $request->merge(['start_date' => $cutOffDateTime]);
+        $request->merge(['end_date' => $today]);
         $logisticRequest = Agency::getList($request, false)->get();
 
         $data = [
@@ -180,16 +181,6 @@ class LogisticRequestController extends Controller
         return $changeStatusParam;
     }
 
-    public function stockCheking(Request $request)
-    {
-        $param = [
-            'applicant_id' => 'required|numeric',
-            'stock_checking_status' => 'required|string'
-        ];
-        $applicant = (Validation::validate($request, $param)) ? $this->updateApplicant($request) : null;
-        return response()->format(Response::HTTP_OK, 'success', $applicant);
-    }
-
     public function masterFaskesCheck($request)
     {
         return $request = (!MasterFaskes::find($request->master_faskes_id)) ? $this->alloableAgencyType($request) : $request;
@@ -235,8 +226,8 @@ class LogisticRequestController extends Controller
             'applicant_file' => 'required|mimes:jpeg,jpg,png,pdf|max:10240'
         ];
         $response = Validation::validate($request, $param);
-        if ($response->getStatusCode() === 200) {
-            $request->request->add(['applicant_id' => $id]);
+        if ($response->getStatusCode() === Response::HTTP_OK) {
+            $request->merge(['applicant_id' => $id]);
             $response = FileUpload::storeApplicantFile($request);
             $applicant = Applicant::where('id', '=', $request->applicant_id)->update(['file' => $response->id]);
             Validation::setCompleteness($request);
