@@ -6,6 +6,7 @@ use App\MasterFaskes;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Validation;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Storage;
 
 class MasterFaskesController extends Controller
@@ -13,18 +14,14 @@ class MasterFaskesController extends Controller
     public function index(Request $request)
     {
         $data = MasterFaskes::getFaskesList($request);
-        $response = response()->format(200, 'success', $data);
+        $response = response()->format(Response::HTTP_OK, 'success', $data);
         return $response;
     }
 
     public function show($id)
     {
-        try {
-            $data =  MasterFaskes::findOrFail($id);
-            return response()->format(200, 'success', $data);
-        } catch (\Exception $exception) {
-            return response()->format(400, $exception->getMessage());
-        }
+        $data =  MasterFaskes::find($id);
+        return response()->format(Response::HTTP_OK, 'success', $data);
     }
 
     public function store(Request $request)
@@ -39,16 +36,16 @@ class MasterFaskesController extends Controller
             'permit_file' => 'required|mimes:jpeg,jpg,png|max:10240'
         ];
         $response = Validation::validate($request, $param);
-        if ($response->getStatusCode() === 200) {
+        if ($response->getStatusCode() === Response::HTTP_OK) {
             try {
                 $model->fill($request->input());
                 $model->verification_status = 'not_verified';
                 $model->is_imported = 0;
                 $model->permit_file = $this->permitLetterStore($request);
                 $model->save();
-                $response = response()->format(200, 'success', $model);
+                $response = response()->format(Response::HTTP_OK, 'success', $model);
             } catch (\Exception $e) {
-                $response = response()->format(400, $e->getMessage());
+                $response = response()->format(Response::HTTP_UNPROCESSABLE_ENTITY, $e->getMessage());
             }
         }
         return $response;
@@ -58,15 +55,15 @@ class MasterFaskesController extends Controller
     {
         $param = ['verification_status' => 'required'];
         $response = Validation::validate($request, $param);
-        if ($response->getStatusCode() === 200) {
+        if ($response->getStatusCode() === Response::HTTP_OK) {
             if ($request->verification_status == 'verified' || $request->verification_status == 'rejected') {
                 try {
                     $model =  MasterFaskes::findOrFail($id);
                     $model->verification_status = $request->verification_status;
                     $model->save();
-                    $response = response()->format(200, 'success', $model);
+                    $response = response()->format(Response::HTTP_OK, 'success', $model);
                 } catch (\Exception $e) {
-                    $response = response()->json(array('message' => 'could_not_verify_faskes'), 500);
+                    $response = response()->json(array('message' => 'could_not_verify_faskes'), Response::HTTP_UNPROCESSABLE_ENTITY);
                 }
             }
         }
@@ -77,7 +74,7 @@ class MasterFaskesController extends Controller
     {
         $path = null;
         if ($request->hasFile('permit_file')) {
-            $path = Storage::disk('s3')->put('registration/letter', $request->permit_file);
+            $path = Storage::put('registration/letter', $request->permit_file);
         }
         return $path;
     }
