@@ -54,26 +54,26 @@ class MasterFaskes extends Model
 
     static function getFaskesList($request)
     {
-        $limit = $request->filled('limit') ? $request->input('limit') : 20;
-        $sort = $request->filled('sort') ? $request->input('sort') : 'asc';
+        $limit = $request->input('limit', 20);
+        $sort = $request->input('sort', 'asc');
 
         $data = self::with('masterFaskesType')
         ->where(function ($query) use ($request) {
-            if ($request->filled('nama_faskes')) {
+            if ($request->has('nama_faskes')) {
                 $query->where('master_faskes.nama_faskes', 'LIKE', "%{$request->input('nama_faskes')}%");
             }
 
-            if ($request->filled('id_tipe_faskes')) {
+            if ($request->has('id_tipe_faskes')) {
                 $query->where('master_faskes.id_tipe_faskes', '=', $request->input('id_tipe_faskes'));
             }
 
-            if ($request->filled('verification_status')) {
+            if ($request->has('verification_status')) {
                 $query->where('master_faskes.verification_status', '=', $request->input('verification_status'));
             } else {
                 $query->where('master_faskes.verification_status', '=', self::STATUS_VERIFIED);
             }
 
-            if ($request->filled('is_imported')) {
+            if ($request->has('is_imported')) {
                 $query->where('master_faskes.is_imported', $request->input('is_imported'));
             }
 
