@@ -13,7 +13,7 @@ return [
     |
     */
 
-    'default' => env('FILESYSTEM_DRIVER', 'local'),
+    'default' => env('FILESYSTEM_DRIVER', 's3'),
 
     /*
     |--------------------------------------------------------------------------
@@ -62,6 +62,12 @@ return [
             'region' => env('AWS_DEFAULT_REGION'),
             'bucket' => env('AWS_BUCKET'),
             'url' => env('AWS_CLOUDFRONT_URL'),
+        ],
+
+        'testing' => [
+            'driver' => 'local',
+            'root' => storage_path('framework/testing/disks/'),
+            'visibility' => 'public',
         ],
 
     ],
