@@ -16,9 +16,14 @@ use Faker\Generator as Faker;
 $factory->define(App\User::class, function (Faker $faker) {
     return [
         'name' => $faker->name,
-        'email' => $faker->unique()->safeEmail,
-        'email_verified_at' => now(),
-        'password' => '$2y$10$TKh8H1.PfQx37YgCzwiKb.KjNyWgaHb9cbcoQgdIVFlYg7B77UdFm', // secret
-        'remember_token' => str_random(10),
+        'username' => $faker->userName,
+        'name_district_city' => $faker->state,
+        'code_district_city' => $faker->numerify('##.##'),
+        'roles' => 'dinkesprov',
+        'agency_name' => 'Dinas Kesehatan',
+        'email' => $faker->email,
+        'password' => bcrypt('secret'), // secret
+        'handphone' => $faker->phoneNumber,
+        'phase' => 'surat',
     ];
 });
