@@ -15,9 +15,9 @@ class UpdateColumnsInMasterFaskesTable extends Migration
     {
         Schema::table('master_faskes', function (Blueprint $table) {
             $table->increments('id')->change();
-            $table->integer('faskes_origin_id')->nullable()->after('id');
-            $table->string('nama_atasan')->after('alamat');
-            $table->string('nomor_registrasi')->after('alamat');
+            $table->integer('faskes_origin_id')->after('id')->nullable();
+            $table->string('nama_atasan')->after('alamat')->nullable();
+            $table->string('nomor_registrasi')->after('alamat')->nullable();
         });
     }
 
